Make public key from hex coords construction public API
